+++ src/pages/Login/Signup.scss
@@ -16,20 +16,19 @@
   font-weight: 900;
   color: #ff0658;
 }
-.Login {
+.Signup {
   position: fixed;
   top: 0;
   right: 0;
   bottom: 0;
   left: 0;
   background: rgba(0, 0, 0, 0.6);
-
-  .underLogin {
+  .underSignup {
     @include flexCenter;
     height: 100%;
     width: 100%;
-
-    .loginBox {
+    
+    .signUpBox {
       @include flexCenter;
       flex-direction: column;
       background-color: #fafafa;
@@ -37,24 +36,52 @@
       width: 21em;
       height: 35em;
 
-      .loginLogo {
+      .signUpLogo {
         display: flex;
         flex-direction: column;
         align-items: center;
 
-        .loginPageLogoHigh {
+        .sigUpPageLogoHigh {
           @include logoDetail;
         }
 
-        .loginPageLogoLow {
+        .sigUpPageLogoLow {
           @include logoDetail;
         }
       }
 
-      .loginText {
+      .signUpText {
         margin: 0.8em 0em;
         padding: 0.8em 0em;
         font-weight: bold;
+      }
+
+      .nameInput {
+        margin: 0.2em 0em;
+        background-color: #f5f5f7;
+        caret-color: #ff0658;
+        @include inputAndBtn;
+      }
+
+      .nameInput::placeholder {
+        color: #cacaca;
+        padding-left: 0.5em;
+      }
+
+      .nameInput:focus {
+        outline: none;
+      }
+
+      .displayNone {
+        display: none;
+      }
+
+      .inputStatus {
+        padding-top: 0.2ex;
+        font-size: 0.4em;
+        width: 88%;
+        display: flex;
+        color: #ff0658;
       }
 
       .emailInput {
@@ -72,19 +99,9 @@
       .emailInput:focus {
         outline: none;
       }
-      .displayNone {
-        display: none;
-      }
 
-      .inputStatus {
-        padding-top: 0.2ex;
-        font-size: 0.4em;
-        width: 88%;
-        display: flex;
-        color: #ff0658;
-      }
       .passwordInput {
-        margin: 0.5em 0em;
+        margin: 0.2em 0em;
         background-color: #f5f5f7;
         caret-color: #ff0658;
         @include inputAndBtn;
@@ -94,11 +111,12 @@
         color: #cacaca;
         padding-left: 0.5em;
       }
+
       .passwordInput:focus {
         outline: none;
       }
-
-      .loginBtn {
+      // }
+      .signUpBtn {
         color: white;
         font-size: 0.8em;
         margin: 1em 0em;
@@ -106,26 +124,18 @@
         @include inputAndBtn;
       }
 
-      .loginBtn:focus {
+      .signUpBtn:focus {
         outline: 0;
       }
 
-      .forgotPwTxt {
-        margin: 0.4em 0em;
-        color: #ff0658;
-        font-weight: 350;
-        font-size: 0.9em;
-        text-decoration: none;
-      }
-
-      .noAccountTxt {
+      .signUpCheckTxt {
         display: flex;
         color: grey;
         margin: 0.7em 0em;
         font-weight: 350;
         font-size: 0.9em;
 
-        .signUpNowBtn {
+        .loginNowBtn {
           padding-left: 0.3em;
           color: #ff0658;
           text-decoration: none;
