+++ src/pages/MainPage/MainPage.scss
@@ -1,5 +1,15 @@
-.MainPage {
+.MainTest{
+    // display:flex;
+    // justify-content: center;
+    // align-items: center;
+    // height: 100vh;
+}
+.mainBody {
     display:flex;
     justify-content: center;
     align-items: center;
+    font-size: 1em;
+    width: 100%;
+    height: 40em;
+    border: 2px solid black;
 }
