--- src/pages/Signup/Signup.scss
@@ -1,144 +1,0 @@
-.Signup {
-  background-color: #b2b2b2;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-
-  .signUpBox {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    background-color: #fafafa;
-    border-radius: 2%;
-    width: 21em;
-    height: 35em;
-
-    .signUpLogo {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-
-      .sigUpPageLogoHigh {
-        font-family: Anton;
-        font-size: 2em;
-        font-style: bold;
-        font-weight: 900;
-        color: #ff0658;
-      }
-
-      .sigUpPageLogoLow {
-        font-family: Anton;
-        font-size: 2em;
-        font-style: bold;
-        font-weight: 900;
-        color: #ff0658;
-      }
-    }
-
-    .signUpText {
-      margin: 0.8em 0em;
-      padding: 0.8em 0em;
-      font-weight: bold;
-    }
-
-    .nameInput {
-      margin: 0.2em 0em;
-      width: 90%;
-      height: 7%;
-      border: 0px;
-      background-color: #f5f5f7;
-      border-radius: 0.3em;
-      caret-color: #ff0658;
-    }
-
-    .nameInput::placeholder {
-      color: #cacaca;
-      padding-left: 0.5em;
-    }
-
-    .nameInput:focus {
-      outline: none;
-    }
-
-    .displayNone {
-      display: none;
-    }
-
-    .inputStatus {
-      padding-top: 0.2ex;
-      font-size: 0.4em;
-      width: 88%;
-      display: flex;
-      color: #ff0658;
-    }
-
-    .emailInput {
-      margin: 0.2em 0em;
-      width: 90%;
-      height: 7%;
-      border: 0px;
-      background-color: #f5f5f7;
-      border-radius: 0.3em;
-      caret-color: #ff0658;
-    }
-
-    .emailInput::placeholder {
-      color: #cacaca;
-      padding-left: 0.5em;
-    }
-
-    .emailInput:focus {
-      outline: none;
-    }
-
-    .passwordInput {
-      margin: 0.2em 0em;
-      width: 90%;
-      height: 7%;
-      border: 0px;
-      background-color: #f5f5f7;
-      border-radius: 0.3em;
-      caret-color: #ff0658;
-    }
-
-    .passwordInput::placeholder {
-      color: #cacaca;
-      padding-left: 0.5em;
-    }
-
-    .passwordInput:focus {
-      outline: none;
-    }
-    // }
-    .signUpBtn {
-      color: white;
-      font-size: 0.8em;
-      margin: 1em 0em;
-      width: 90%;
-      height: 7%;
-      border: 0px;
-      background-color: #ff0658;
-      border-radius: 0.3em;
-    }
-
-    .signUpBtn:focus {
-      outline: 0;
-    }
-
-    .signUpCheckTxt {
-      display: flex;
-      color: grey;
-      margin: 0.7em 0em;
-      font-weight: 350;
-      font-size: 0.9em;
-
-      .loginNowBtn {
-        padding-left: 0.3em;
-        color: #ff0658;
-        text-decoration: none;
-      }
-    }
-  }
-}
